Move run command to separate package
